user_registration/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
databases==0.9.0
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.2.1
h11==0.14.0
idna==3.10
numpy==2.2.2
pandas==2.2.3
psycopg2==2.9.10
pyasn1==0.4.8
pydantic==2.11.3
pydantic_core==2.33.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2024.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.2
